# improved performance of merge() utils